pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "groupme-travel-bot"
version = "0.1.0"
description = "A GroupMe chatbot for collaborative travel planning"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
    "azure-cosmos>=4.15.0",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "weasyprint>=68.1",
    "jinja2>=3.1.6",
    "azure-identity>=1.25.3",
    "azure-storage-blob>=12.28.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
    "httpx>=0.28.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
